refactor(links): extract LinkInfoEditor (#416)
